Refactor first days
Mainly package and class naming
