+++ src/main/kotlin/me/quido/Main.kt
@@ -2,18 +2,18 @@ package me.quido
 
 import me.quido.util.Solver
 import me.quido.util.readFile
-import me.quido.year2023.calibrating.Trebuchet
-import me.quido.year2023.engineering.Gondola
-import me.quido.year2023.gaming.ColoredCubes
-import me.quido.year2023.gaming.ScratchCards
+import me.quido.year2023.engineering.TrebuchetCalibrator
+import me.quido.year2023.engineering.EngineReader
+import me.quido.year2023.gaming.CubeCalculator
+import me.quido.year2023.gaming.ScratchCardsCalculator
 import java.time.OffsetDateTime
 import java.time.temporal.ChronoUnit.MILLIS
 
 private val solvers2023: List<Solver> = listOf(
-    Trebuchet(),
-    ColoredCubes(),
-    Gondola(),
-    ScratchCards(),
+    TrebuchetCalibrator(),
+    CubeCalculator(),
+    EngineReader(),
+    ScratchCardsCalculator(),
 )
